Make newest kills appear at the top, work on fake kills
I could not solve the issue where on the page load, its not scrolled to the bottom. We do that so it reads like a chat channel, or kill feed in a game, with the newest kills at the bottom.

Despite trying various things, the async network requests or CSS would pop in after we scroll to the bottom, so when the page loaded it would only be scrolled down about halfway or so. Couldn't figure out a good solution, despite this working fine on TEAW's chat page without issue (its basically the same code idk why it  doesn't work).